chore(connlib): move IpProvider to tunnel
